revert change to rounding in plan_buffer_line This was a needless change since the arguments were float and the returned value is a long anyway.

+++ src/planner.c
@@ -367,8 +367,7 @@ uint8_t plan_buffer_line(float *target, plan_line_data_t *pl_data)
         delta_mm = (target_steps[idx] - position_steps[idx])/settings.steps_per_mm[idx];
       }
     #else
-      //target_steps[idx] = lroundf(target[idx]*settings.steps_per_mm[idx]); Paul; replace roundf 16 bits with round 32 bits
-      target_steps[idx] = lround(target[idx]*settings.steps_per_mm[idx]);
+      target_steps[idx] = lroundf(target[idx]*settings.steps_per_mm[idx]);
       block->steps[idx] = abs(target_steps[idx]-position_steps[idx]);
       block->step_event_count = max(block->step_event_count, block->steps[idx]);
       delta_mm = (target_steps[idx] - position_steps[idx])/settings.steps_per_mm[idx];
